requirements.txt
# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Database
databricks-sql-connector>=3.0.0

# Caching and performance
redis>=5.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Validation
pydantic>=2.0.0

# Async support
aiohttp>=3.9.0

